:label: add types:make type file
typesファイルを作成しそこにまとめた
